Move Delete button to Directory Detail page
