+++ src/pages/directory-detail/directory-detail.html
@@ -8,8 +8,11 @@
     <img class="profile-photo" src="{{profile?.photoUrl ? transformUrl(profile.photoUrl) : 'assets/no_photo.png'}}" />
     <ion-item class="photo">
       <h1>{{ profile?.firstName }} {{ profile?.lastName }}</h1>
-      <button *ngIf="isAdmin" ion-button (click)="goToProfileAdminScreen(profile.id)">
+      <button *ngIf="isAdmin" ion-button (click)="goToProfileAdminScreen(profile.id)" color="primary">
         Edit Profile
+      </button>
+      <button *ngIf="isAdmin" ion-button (click)="deleteProfile(profile.id)" color="danger">
+        Delete Profile
       </button>
     </ion-item>
   </ion-card>
