+++ src/pages/directory-detail/directory-detail.scss
@@ -1,3 +1,7 @@
 .profile-photo {
   max-height: 200px;
 }
+
+.delete-confirmation {
+  color: #f53d3d !important;
+}
