--- src/pages/directory/directory.scss
@@ -1,3 +1,0 @@
-.delete-confirmation {
-  color: #f53d3d !important;
-}
